backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-analytics-data
google-api-python-client
httpx
